[*] Switch to single file as per original author's request; [setup.py] Repurpose for single file
