+++ docopt_c.py
@@ -25,26 +25,401 @@ Arguments:
 
 """
 
-from __future__ import print_function
+__author__ = "Vladimir Keleshev"
+__version__ = "2.0rc2"
+__description__ = "C generator for language for description of command-line interfaces"
 
 import numbers
 import os.path
 import re
 import textwrap
-from functools import partial
 from string import Template
 
 import sys
 
-from pkg_resources import resource_filename
+import docopt
 
-from docopt_c import docopt
+template_h = """
+#ifndef DOCOPT_$header_no_ext_H
+#define DOCOPT_$header_no_ext_H
 
-_data_join = partial(
-    os.path.join,
-    os.path.dirname(resource_filename(sys.modules[__name__].__name__.partition(".")[2], "__init__.py")),
-    "_data"
-)
+#include <stddef.h>
+
+#if defined(__STDC__) && defined(__STDC_VERSION__) && __STDC_VERSION__ >= 199901L
+
+#include <stdbool.h>
+
+#else
+
+#ifdef true
+#undef true
+#endif
+#ifdef false
+#undef false
+#endif
+#ifdef bool
+#undef bool
+#endif
+
+#define true 1
+#define false !true
+typedef int bool;
+
+#endif
+
+#if defined(_AIX)
+
+#include <sys/limits.h>
+
+#elif defined(__FreeBSD__) || defined(__NetBSD__)
+|| defined(__OpenBSD__) || defined(__bsdi__)
+|| defined(__DragonFly__) || defined(macintosh)
+|| defined(__APPLE__) || defined(__APPLE_CC__)
+
+#include <sys/syslimits.h>
+
+#elif defined(__HAIKU__)
+
+#include <system/user_runtime.h>
+
+#elif defined(__linux__) || defined(linux) || defined(__linux)
+
+#include <linux/version.h>
+
+#if LINUX_VERSION_CODE <= KERNEL_VERSION(2,6,22)
+
+#include <linux/limits.h>
+
+#else
+
+#define ARG_MAX       131072    /* # bytes of args + environ for exec() */
+/* it's no longer defined, see this example and more at https://unix.stackexchange.com/q/120642 */
+
+#endif
+
+#elif (defined(__FreeBSD__) || defined(__NetBSD__) || defined(__OpenBSD__) || defined(__bsdi__) \
+ || defined(__DragonFly__) || defined(macintosh) || defined(__APPLE__) || defined(__APPLE_CC__))
+
+#include <sys/param.h>
+
+#else
+
+#include <limits.h>
+
+#endif
+
+#ifndef ARG_MAX
+#ifdef NCARGS
+#define ARG_MAX NCARGS
+#else
+#define ARG_MAX 131072
+#endif
+#endif
+
+struct DocoptArgs {
+    $commands$arguments$flags$options
+    /* special */
+    const char *usage_pattern;
+    const char *help_message[$help_message_n];
+};
+
+struct DocoptArgs docopt(int, char *[], bool, const char *);
+
+#endif
+"""
+
+template_c = """
+#include <stdio.h>
+#include <stdlib.h>
+#include <string.h>
+
+#include "$header_name"
+
+struct Command {
+    const char *name;
+    bool value;
+};
+
+struct Argument {
+    const char *name;
+    const char *value;
+    const char *array[ARG_MAX];
+};
+
+struct Option {
+    const char *oshort;
+    const char *olong;
+    bool argcount;
+    bool value;
+    const char *argument;
+};
+
+struct Elements {
+    int n_commands;
+    int n_arguments;
+    int n_options;
+    struct Command *commands;
+    struct Argument *arguments;
+    struct Option *options;
+};
+
+
+/*
+ * Tokens object
+ */
+
+struct Tokens {
+    int argc;
+    char **argv;
+    int i;
+    char *current;
+};
+
+const char usage_pattern[] =
+        $usage_pattern;
+
+struct Tokens tokens_new(int argc, char **argv) {
+    struct Tokens ts;
+    ts.argc = argc;
+    ts.argv = argv;
+    ts.i = 0;
+    ts.current = argv[0];
+    return ts;
+}
+
+struct Tokens *tokens_move(struct Tokens *ts) {
+    if (ts->i < ts->argc) {
+        ts->current = ts->argv[++ts->i];
+    }
+    if (ts->i == ts->argc) {
+        ts->current = NULL;
+    }
+    return ts;
+}
+
+
+/*
+ * ARGV parsing functions
+ */
+
+int parse_doubledash(struct Tokens *ts, struct Elements *elements) {
+    /*
+    int n_commands = elements->n_commands;
+    int n_arguments = elements->n_arguments;
+    Command *commands = elements->commands;
+    Argument *arguments = elements->arguments;
+
+    not implemented yet
+    return parsed + [Argument(None, v) for v in tokens]
+    */
+    return 0;
+}
+
+int parse_long(struct Tokens *ts, struct Elements *elements) {
+    int i;
+    int len_prefix;
+    int n_options = elements->n_options;
+    char *eq = strchr(ts->current, '=');
+    struct Option *option;
+    struct Option *options = elements->options;
+
+    len_prefix = (eq - (ts->current)) / sizeof(char);
+    for (i = 0; i < n_options; i++) {
+        option = &options[i];
+        if (!strncmp(ts->current, option->olong, len_prefix))
+            break;
+    }
+    if (i == n_options) {
+        /* TODO: %s is not a unique prefix */
+        fprintf(stderr, "%s is not recognized\n", ts->current);
+        return 1;
+    }
+    tokens_move(ts);
+    if (option->argcount) {
+        if (eq == NULL) {
+            if (ts->current == NULL) {
+                fprintf(stderr, "%s requires argument\n", option->olong);
+                return 1;
+            }
+            option->argument = ts->current;
+            tokens_move(ts);
+        } else {
+            option->argument = eq + 1;
+        }
+    } else {
+        if (eq != NULL) {
+            fprintf(stderr, "%s must not have an argument\n", option->olong);
+            return 1;
+        }
+        option->value = true;
+    }
+    return 0;
+}
+
+int parse_shorts(struct Tokens *ts, struct Elements *elements) {
+    char *raw;
+    int i;
+    int n_options = elements->n_options;
+    struct Option *option;
+    struct Option *options = elements->options;
+
+    raw = &ts->current[1];
+    tokens_move(ts);
+    while (raw[0] != '\0') {
+        for (i = 0; i < n_options; i++) {
+            option = &options[i];
+            if (option->oshort != NULL && option->oshort[1] == raw[0])
+                break;
+        }
+        if (i == n_options) {
+            /* TODO -%s is specified ambiguously %d times */
+            fprintf(stderr, "-%c is not recognized\n", raw[0]);
+            return EXIT_FAILURE;
+        }
+        raw++;
+        if (!option->argcount) {
+            option->value = true;
+        } else {
+            if (raw[0] == '\0') {
+                if (ts->current == NULL) {
+                    fprintf(stderr, "%s requires argument\n", option->oshort);
+                    return EXIT_FAILURE;
+                }
+                raw = ts->current;
+                tokens_move(ts);
+            }
+            option->argument = raw;
+            break;
+        }
+    }
+    return EXIT_SUCCESS;
+}
+
+int parse_argcmd(struct Tokens *ts, struct Elements *elements) {
+    int i;
+    int n_commands = elements->n_commands;
+    /* int n_arguments = elements->n_arguments; */
+    struct Command *command;
+    struct Command *commands = elements->commands;
+    /* Argument *arguments = elements->arguments; */
+
+    for (i = 0; i < n_commands; i++) {
+        command = &commands[i];
+        if (strcmp(command->name, ts->current) == 0) {
+            command->value = true;
+            tokens_move(ts);
+            return EXIT_SUCCESS;
+        }
+    }
+    /* not implemented yet, just skip for now
+       parsed.append(Argument(None, tokens.move())) */
+    /*
+    fprintf(stderr, "! argument '%s' has been ignored\n", ts->current);
+    fprintf(stderr, "  '");
+    for (i=0; i<ts->argc ; i++)
+        fprintf(stderr, "%s ", ts->argv[i]);
+    fprintf(stderr, "'\n");
+    */
+    tokens_move(ts);
+    return EXIT_SUCCESS;
+}
+
+int parse_args(struct Tokens *ts, struct Elements *elements) {
+    int ret = EXIT_FAILURE;
+
+    while (ts->current != NULL) {
+        if (strcmp(ts->current, "--") == 0) {
+            ret = parse_doubledash(ts, elements);
+            if (ret == EXIT_FAILURE) break;
+        } else if (ts->current[0] == '-' && ts->current[1] == '-') {
+            ret = parse_long(ts, elements);
+        } else if (ts->current[0] == '-' && ts->current[1] != '\0') {
+            ret = parse_shorts(ts, elements);
+        } else
+            ret = parse_argcmd(ts, elements);
+        if (ret) return ret;
+    }
+    return ret;
+}
+
+int elems_to_args(struct Elements *elements, struct DocoptArgs *args,
+                     const bool help, const char *version) {
+    struct Command *command;
+    struct Argument *argument;
+    struct Option *option;
+    int i, j;
+
+    /* fix gcc-related compiler warnings (unused) */
+    (void) command;
+    (void) argument;
+
+    /* options */
+    for (i = 0; i < elements->n_options; i++) {
+        option = &elements->options[i];
+        if (help && option->value && strcmp(option->olong, "--help") == 0) {
+            for (j = 0; j < $help_message_n; j++)
+                puts(args->help_message[j]);
+            return EXIT_FAILURE;
+        } else if (version && option->value &&
+                   strcmp(option->olong, "--version") == 0) {
+            puts(version);
+            return EXIT_FAILURE;
+        }$if_flag$if_option
+    }
+    /* commands */
+    for (i = 0; i < elements->n_commands; i++) {
+        command = &elements->commands[i];
+        $if_command
+    }
+    /* arguments */
+    for (i = 0; i < elements->n_arguments; i++) {
+        argument = &elements->arguments[i];
+        $if_argument
+    }
+    return EXIT_SUCCESS;
+}
+
+
+/*
+ * Main docopt function
+ */
+
+struct DocoptArgs docopt(int argc, char *argv[], const bool help, const char *version) {
+    struct DocoptArgs args = {$defaults
+            usage_pattern,
+            $help_message
+    };
+    struct Tokens ts;
+    struct Command commands[] = {$elems_cmds
+    };
+    struct Argument arguments[] = {$elems_args
+    };
+    struct Option options[] = {$elems_opts
+    };
+    struct Elements elements;
+    int return_code = EXIT_SUCCESS;
+
+    elements.n_commands = $t_elems_n_commands;
+    elements.n_arguments = $t_elems_n_arguments;
+    elements.n_options = $t_elems_n_options;
+    elements.commands = commands;
+    elements.arguments = arguments;
+    elements.options = options;
+
+    if (argc == 1) {
+        argv[argc++] = "--help";
+        argv[argc++] = NULL;
+        return_code = EXIT_FAILURE;
+    }
+
+    ts = tokens_new(argc, argv);
+    if (parse_args(&ts, &elements))
+        exit(EXIT_FAILURE);
+    if (elems_to_args(&elements, &args, help, version))
+        exit(return_code);
+    return args;
+}
+"""
 
 def to_initializer(val):
     if isinstance(val, (str, type(None), bool, numbers.Number)):
@@ -170,15 +545,15 @@ def main():
         else:
             args['<docopt>'] = sys.stdin.read()
         if args['--template'] is None:
-            args['--template'] = os.path.join(
-                os.path.dirname(os.path.realpath(__file__)), _data_join("template.c"))
+            args['--template'] = template_c
+        else:
+            with open(args['--template'], 'rt') as f:
+                args['--template'] = f.read()
         if args['--template-header'] is None:
-            args['--template-header'] = os.path.join(
-                os.path.dirname(args['--template']), _data_join("template.h"))
-        with open(args['--template'], 'rt') as f:
-            args['--template'] = f.read()
-        with open(args['--template-header'], 'rt') as f:
-            args['--template-header'] = f.read()
+            args['--template-header'] = template_h
+        else:
+            with open(args['--template-header'], 'rt') as f:
+                args['--template-header'] = f.read()
     except IOError as e:
         sys.exit(e)
 
@@ -256,11 +631,15 @@ def main():
         for opt in options
     )
 
-    if args['--output-name'].endswith('.c'):
-        header_output_name = os.path.splitext(args['--output-name'])[0] + '.h'
+    if not args['--output-name']:
+        header_output_name = '<stdout>'
     else:
-        args['--output-name'] = args['--output-name'] + '.h'
-        header_output_name = args['--output-name'] + '.c'
+        base, ext = os.path.splitext(args['--output-name'])
+        if ext not in frozenset(('.h', '.c')):
+            base = args['--output-name']
+
+        args['--output-name'] = "{base}.c".format(base=base)
+        header_output_name = "{base}.h".format(base=base)
 
     header_name = os.path.basename(header_output_name)
 
@@ -298,10 +677,10 @@ def main():
         print(template_out.strip(), '\n')
     else:
         try:
-            with open(args['--output-name'], 'w') as f:
+            with open(sys.stdout if args['--output-name'] in (None, "<stdout>") else args['--output-name'], 'w') as f:
                 f.write(template_out.strip() + '\n')
 
-            with open(header_output_name, 'w') as f:
+            with open(sys.stdout if header_output_name == "<stdout>" else header_output_name, 'w') as f:
                 f.write(template_header_out.strip() + '\n')
 
         except IOError as e:
