--- docopt_c/__init__.py
@@ -1,5 +1,0 @@
-# -*- coding: utf-8 -*-
-
-__author__ = "Vladimir Keleshev"
-__version__ = "2.0.0-rc2"
-__description__ = "C generator for language for description of command-line interfaces"
