+++ test/example.docopt
